Re-organizing lib folder.  Fleshing out example
