+++ example/ios/Runner/Info.plist
@@ -2,14 +2,10 @@
 <!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
 <plist version="1.0">
 <dict>
-	<key>NSMotionUsageDescription</key>
-	<string>Motion always</string>
-	<key>NSLocationWhenInUseUsageDescription</key>
-	<string>When in use</string>
-	<key>NSLocationAlwaysAndWhenInUseUsageDescription</key>
-	<string>Always and When in Use</string>
 	<key>CFBundleDevelopmentRegion</key>
 	<string>en</string>
+	<key>CFBundleDisplayName</key>
+	<string>flutter_background_geolocation_example</string>
 	<key>CFBundleExecutable</key>
 	<string>$(EXECUTABLE_NAME)</string>
 	<key>CFBundleIdentifier</key>
@@ -28,6 +24,12 @@
 	<string>1</string>
 	<key>LSRequiresIPhoneOS</key>
 	<true/>
+	<key>NSLocationAlwaysAndWhenInUseUsageDescription</key>
+	<string>Always and When in Use</string>
+	<key>NSLocationWhenInUseUsageDescription</key>
+	<string>When in use</string>
+	<key>NSMotionUsageDescription</key>
+	<string>Motion always</string>
 	<key>UIBackgroundModes</key>
 	<array>
 		<string>audio</string>
