+++ example/lib/constants.dart
@@ -1,0 +1,4 @@
+part of flt_background_geolocation_example;
+
+const COLOR_GOLD = Color.fromRGBO(254, 221, 30, 1.0);
+const COLOR_GREEN = Color.fromRGBO(17, 183, 0, 1.0);
