+++ example/pubspec.yaml
@@ -4,6 +4,10 @@ description: Demonstrates how to use the flutter_background_geolocation plugin.
 dependencies:
   flutter:
     sdk: flutter
+  google_maps_flutter:
+     git:
+       url: git://github.com/flutter/plugins
+       path: packages/google_maps_flutter
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
