+++ lib/src/models/activity_change_event.dart
