+++ lib/src/models/connectivity_change_event.dart
