+++ lib/src/models/geofence.dart
