+++ lib/src/models/geofence_event.dart
