+++ lib/src/models/heartbeat_event.dart
