+++ lib/src/models/http_event.dart
