+++ lib/src/models/power_save_change_event.dart
